refactor: 도메인에 인터페이스를 두고 Jgrapht 의존성 분리
