+++ src/main/java/subway/domain/path/Path.java
@@ -1,24 +1,10 @@
 package subway.domain.path;
 
 import java.util.List;
-import subway.domain.core.Section;
 
-public class Path {
+public interface Path {
 
-    private final List<SectionEdge> sectionEdges;
+    int calculateTotalDistance();
 
-    public Path(final List<SectionEdge> sectionEdges) {
-        this.sectionEdges = sectionEdges;
-    }
-
-    public int calculateTotalDistance() {
-        return sectionEdges.stream()
-                .map(SectionEdge::toSection)
-                .mapToInt(Section::getDistanceValue)
-                .sum();
-    }
-
-    public List<SectionEdge> getSectionEdges() {
-        return sectionEdges;
-    }
+    List<SectionEdge> getSectionEdges();
 }
