+++ src/test/java/subway/domain/fare/DistanceFarePolicyTest.java
@@ -1,14 +1,13 @@
 package subway.domain.fare;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.mock;
 
-import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import subway.domain.core.Section;
 import subway.domain.path.Path;
-import subway.domain.path.SectionEdge;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
@@ -19,10 +18,8 @@ class DistanceFarePolicyTest {
     @Test
     void 입력받은_거리가_10키로_이하인_경우_추가운임이_발생하지_않는다() {
         // given
-        final Path path = new Path(List.of(
-                new SectionEdge(new Section("A", "B", 4), 100, 1),
-                new SectionEdge(new Section("B", "C", 5), 100, 1)
-        ));
+        final Path path = mock(Path.class);
+        given(path.calculateTotalDistance()).willReturn(10);
         final Passenger passenger = new Passenger(20);
 
         // when
@@ -35,10 +32,8 @@ class DistanceFarePolicyTest {
     @Test
     void 입력받은_거리가_10키로를_초과하고_50키로_이하인_경우_추가운임이_5KM당_100원이_발생한다() {
         // given
-        final Path path = new Path(List.of(
-                new SectionEdge(new Section("A", "B", 20), 100, 1),
-                new SectionEdge(new Section("B", "C", 7), 100, 1)
-        ));
+        final Path path = mock(Path.class);
+        given(path.calculateTotalDistance()).willReturn(27);
         final Passenger passenger = new Passenger(20);
 
         // when
@@ -51,10 +46,8 @@ class DistanceFarePolicyTest {
     @Test
     void 입력받은_거리가_50키로를_초과하는_경우_추가운임이_8KM당_100원이_발생한다() {
         // given
-        final Path path = new Path(List.of(
-                new SectionEdge(new Section("A", "B", 60), 100, 1),
-                new SectionEdge(new Section("B", "C", 1), 100, 1)
-        ));
+        final Path path = mock(Path.class);
+        given(path.calculateTotalDistance()).willReturn(61);
         final Passenger passenger = new Passenger(20);
 
         // when
