+++ src/test/java/subway/domain/fare/SubwayFarePolicyTest.java
@@ -1,6 +1,8 @@
 package subway.domain.fare;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.mock;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -23,11 +25,12 @@ class SubwayFarePolicyTest {
     @Test
     void 최종_운임을_계산한다() {
         // given
-        final Path path = new Path(List.of(
-                new SectionEdge(new Section("A", "B", 5), 300, 1),
-                new SectionEdge(new Section("B", "C", 10), 300, 1),
-                new SectionEdge(new Section("C", "T", 10), 500, 2),
-                new SectionEdge(new Section("T", "D", 10), 300, 1)
+        final Path path = mock(Path.class);
+        given(path.calculateTotalDistance()).willReturn(35);
+        given(path.getSectionEdges()).willReturn(List.of(
+                generateSectionEdgeStub(100),
+                generateSectionEdgeStub(500),
+                generateSectionEdgeStub(100)
         ));
         final Passenger passenger = new Passenger(17);
 
@@ -36,5 +39,24 @@ class SubwayFarePolicyTest {
 
         // then
         assertThat(result).isEqualTo(1520);
+    }
+
+    private SectionEdge generateSectionEdgeStub(final int surcharge) {
+        return new SectionEdge() {
+            @Override
+            public Section toSection() {
+                return null;
+            }
+
+            @Override
+            public int getSurcharge() {
+                return surcharge;
+            }
+
+            @Override
+            public long getLineId() {
+                return 0;
+            }
+        };
     }
 }
