+++ src/test/java/subway/jgrapht/JgraphtSectionEdgeTest.java
@@ -1,4 +1,4 @@
-package subway.domain.path;
+package subway.jgrapht;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,15 +6,17 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import subway.domain.core.Section;
+import subway.domain.path.SectionEdge;
+import subway.jgraph.JgraphtSectionEdge;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
-class SectionEdgeTest {
+class JgraphtSectionEdgeTest {
 
     @Test
     void Section을_반환한다() {
         // given
-        final SectionEdge sectionEdge = new SectionEdge(new Section("A", "B", 5), 500, 1);
+        final SectionEdge sectionEdge = new JgraphtSectionEdge(new Section("A", "B", 5), 500, 1);
 
         // when
         final Section result = sectionEdge.toSection();
